# src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Swagger for Plask task API",
        "description": "C.R.C 프로젝트 Swagger",
        "license": {
            "name": "MIT",
            "url": "http://10.120.75.224:3000"
        }
    },
    "host": "10.120.75.224:3000",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "회원가입",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "회원가입시 필요 데이터 수집",
                        "schema": {
                            "$ref": "#/definitions/RegisterData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Register Success"
                    },
                    "404": {
                        "description": "Failed. Bad Request"
                    }
                }
            }
        },
...
        "/suggest/check": {
            "get": {
                "tags": [
                    "suggest"
                ],
                "summary": "게시판 목록 확인",
                "parameters": [
                    {
                        "in": "header",
                        "name": "Token",
                        "type":"string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "suggest Success"
                    },
                    "404": {
                        "description": "Failed. Bad Request"
                    }
                }
            }
        },
        "/suggest/register": {
            "post": {
                "tags": [
                    "suggest"
                ],
                "summary": "게시판 작성",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "suggest Write",
                        "schema": {
                            "$ref": "#/definitions/Suggest_Write"
                        }
                    },
                    {
                        "in": "header",
                        "name": "Token",
                        "type":"string"
                    }
                ],
                
                "responses": {
                    "201": {
                        "description": "suggest Success"
                    },
                    "404": {
                        "description": "Failed. Bad Request"
                    }
                }
            }
        },
        "/suggest/empathy": {
            "post": {
                "tags": [
                    "suggest"
                ],
                "summary": "게시판 공감",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "suggest empathy",
                        "schema": {
                            "$ref": "#/definitions/Suggest_empathy"
                        }
                    },
                    {
                        "in": "header",
                        "name": "Token",
                        "type":"string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "suggest Success"
                    },
                    "404": {
                        "description": "Failed. Bad Request"
                    }
                }
            }
        },
        "/suggest/reply": {
            "post": {
                "tags": [
                    "suggest"
                ],
                "summary": "게시판 응답",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "suggest reply",
                        "schema": {
                            "$ref": "#/definitions/Suggest_reply"
                        }
                    },
                    {
                        "in": "header",
                        "name": "Token",
                        "type":"string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "suggest Success"
                    },
                    "404": {
                        "description": "Failed. Bad Request"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "RegisterData": {
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "student_data": {
                    "type": "string"
                }
            }
        },
...
        "Suggest_Write": {
            "properties": {
                "title": {
                    "type":"string"
                },
                "content": {
                    "type":"string"
                },
                "nickname": {
                    "type":"string"
                },
                "when":{
                    "type":"string"
                }
            }
        },
        "Suggest_empathy": {
            "properties": {
                "suggestid": {
                    "type":"integer"
                }
            }
        },
        "Suggest_reply": {
            "properties": {
                "reply": {
                    "type":"string"
                },
                "suggestid": {
                    "type":"integer"
                }
            }
        },
...
    }
}